# openapi.yaml
openapi: 3.1.0
info:
  title: Collaborative Travel Planner API
  description: |-
    This is the API for the Collaborative Travel Planner application. It allows users to create, manage, and share travel itineraries. The API supports real-time collaboration features through WebSockets.
    Authentication is handled via JWT Bearer tokens obtained from Google OAuth.
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /trips/{tripId}/timeline:
    get:
      summary: Get timeline data for Gantt chart visualization
      description: Retrieves structured timeline data suitable for Gantt chart rendering and project visualization.
      tags:
        - Visualization
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: granularity
          in: query
          description: Time granularity for the timeline
          schema:
            type: string
            enum: [hour, day, week]
            default: day
      responses:
        '200':
          description: Timeline data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineData'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trips/{tripId}/map-data:
    get:
      summary: Get map visualization data
      description: Retrieves structured data for map visualization including markers, routes, and geographic information.
      tags:
        - Visualization
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: includeRoutes
          in: query
          description: Whether to include route calculations
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Map data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MapVisualizationData'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Service not found"
        message:
          type: string
          description: Detailed error description
          example: "The requested service could not be found"
      required:
        - error
    # AI Agent Schemas
...
    TimelineData:
      type: object
      properties:
        tripId:
          type: string
          format: uuid
        timeframe:
          type: object
          properties:
            start:
              type: string
              format: date-time
            end:
              type: string
              format: date-time
            duration:
              type: integer
              description: Duration in days
        tasks:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              type:
                type: string
                enum: [travel, accommodation, activity, meal, free_time]
              start:
                type: string
                format: date-time
              end:
                type: string
                format: date-time
              duration:
                type: integer
                description: Duration in minutes
              dependencies:
                type: array
                items:
                  type: string
                  description: IDs of dependent tasks
              status:
                type: string
                enum: [planned, confirmed, in_progress, completed, cancelled]
              progress:
                type: integer
                minimum: 0
                maximum: 100
              location:
                type: object
                properties:
                  name:
                    type: string
                  coordinates:
                    type: array
                    items:
                      type: number
              cost:
                type: number
              currency:
                type: string
              assignedTo:
                type: array
                items:
                  type: string
                  format: uuid
              priority:
                type: string
                enum: [low, medium, high]
              tags:
                type: array
                items:
                  type: string
        milestones:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              date:
                type: string
                format: date-time
              type:
                type: string
                enum: [departure, arrival, checkpoint, deadline]
              status:
                type: string
                enum: [upcoming, reached, missed]
        criticalPath:
          type: array
          items:
            type: string
            description: Task IDs that form the critical path
        generatedAt:
          type: string
          format: date-time
    MapVisualizationData:
      type: object
      properties:
        tripId:
          type: string
          format: uuid
        bounds:
          type: object
          properties:
            northeast:
              type: object
              properties:
                lat:
                  type: number
                lng:
                  type: number
            southwest:
              type: object
              properties:
                lat:
                  type: number
                lng:
                  type: number
        markers:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              type:
                type: string
                enum: [destination, accommodation, activity, restaurant, transport_hub, custom]
              coordinates:
                type: object
                properties:
                  lat:
                    type: number
                  lng:
                    type: number
              title:
                type: string
              description:
                type: string
              date:
                type: string
                format: date-time
              status:
                type: string
                enum: [planned, confirmed, visited, cancelled]
              metadata:
                type: object
                properties:
                  address:
                    type: string
                  rating:
                    type: number
                  priceLevel:
                    type: integer
                  category:
                    type: string
                  photos:
                    type: array
                    items:
                      type: string
              clustering:
                type: object
                properties:
                  groupId:
                    type: string
                  groupName:
                    type: string
        routes:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              waypoints:
                type: array
                items:
                  type: object
                  properties:
                    lat:
                      type: number
                    lng:
                      type: number
                    name:
                      type: string
                    order:
                      type: integer
              travelMode:
                type: string
                enum: [driving, walking, transit, cycling]
              distance:
                type: number
                description: Distance in meters
              duration:
                type: number
                description: Duration in seconds
              polyline:
                type: string
                description: Encoded polyline for route visualization
              date:
                type: string
                format: date
              status:
                type: string
                enum: [planned, active, completed]
        heatmapData:
          type: array
          items:
            type: object
            properties:
              lat:
                type: number
              lng:
                type: number
              weight:
                type: number
                description: Activity intensity at this location
        layers:
          type: object
          properties:
            showTransport:
              type: boolean
            showAccommodation:
              type: boolean
            showActivities:
              type: boolean
            showRestaurants:
              type: boolean
            showRoutes:
              type: boolean
            showHeatmap:
              type: boolean
        generatedAt:
          type: string
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT